requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=/home/alek/case/projects/sutra_central/sutta_processor/requirements.txt
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, pytest, sutta_processor (setup.py)
autoflake==1.3.1          # via sutta_processor (setup.py)
black==19.10b0            # via sutta_processor (setup.py)
click==7.1.2              # via black, pip-tools
frozendict==1.2           # via sutta_processor (setup.py)
importlib-metadata==1.6.0  # via pluggy, pytest
isort==4.3.21             # via sutta_processor (setup.py)
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pathspec==0.8.0           # via black
pip-tools==5.1.0          # via sutta_processor (setup.py)
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via sutta_processor (setup.py)
pyflakes==2.2.0           # via autoflake
pyparsing==2.4.7          # via packaging
pytest-pythonpath==0.7.3  # via sutta_processor (setup.py)
pytest==5.4.1             # via pytest-pythonpath, sutta_processor (setup.py)
regex==2020.4.4           # via black
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via sutta_processor (setup.py)
six==1.14.0               # via packaging, pip-tools
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
wcwidth==0.1.9            # via pytest
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
